requirements.txt
matplotlib==3.7.1
numpy==1.24.2
pandas==1.5.3
Pillow==9.4.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
kaggle
wandb 
